requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
redis>=5.0.0
httpx>=0.27.0
PyJWT>=2.8.0
groq>=0.12.0
sse-starlette>=2.0.0
strawberry-graphql[fastapi]>=0.240.0
